Retire competition scripts from 2024 and remove usage of --arrays-weak-equiv (#11037)
This pull request creates a copy of the competition scripts to document
the state of the scripts as they were used for the 2024 competition.

It also removes all uses of the experimental option
`--arrays-weak-equiv`. This is done in two commits. The first commit
removes it from the single query script, and the second commit removes
it from the unsat-cores script after the 2024 copies are created.
The problem with `--arrays-weak-equiv` is documented here.