Add conflict processor module (#10738)
This module can be used to minimize explanations of theory lemmas and
conflicts. In practice, around 30% of lemmas can be minimized using this
module. This leads to significant improvements in performance in some
industrial benchmarks.

A followup PR will optionally enable this technique (disabled by
default, as it gives mixed performance on SMT-LIB).

Another use case will be for automatic proof reconstruction, as this
module can infer proofs for theory lemmas in a generic way.