Revert to fresh binder semantics in parser (#11076)
Fixes https://github.com/cvc5/cvc5/issues/11066.

This reverts a change to how variables are parsed as introduced in
https://github.com/cvc5/cvc5/pull/10556.

After this PR, the reverted behavior is available with
`--no-fresh-binders`, but still suffers from potential unsoundness
demonstrated on https://github.com/cvc5/cvc5/issues/11066.

In particular, with non-fresh binders, it is possible to use SMT-LIB
`let` to generate a term that is not properly interpreted by our parser,
as demonstrated by the minimal example on that issue. In particular,
since we do not interpret `let` as object syntax, the minimized
benchmark on that issue leads to a substitution that is not capture
avoiding.

The option `--no-fresh-binders` should be considered moving forward for
proofs / distributed solving but will need a workaround to address the
above issue, either by making `let` object syntax or further restricting
which inputs we allow in the parser.