Restrict the combination of let terms and variable shadowing (#11079)
This reenables non-fresh binder semantics by default (i.e.
`--fresh-binders` is `false`), as introduced in
https://github.com/cvc5/cvc5/pull/10556 and disabled in
https://github.com/cvc5/cvc5/pull/11076.

The motivation for this change is that ALF proofs involving quantifiers
are incorrect on main, and distributed solving is expected to suffer
major performance issues unless `--fresh-binders` is false. This PR is a
medium term solution until we either support object syntax for let (e.g.
a `Kind::LET`) or refactor how these features are handled.

We now restrict the combination of variable shadowing and `let` within
quantified formula bodies to avoid potential for unsoundness.

This means that we are *not* SMT-LIB compliant in that we reject corner
cases like the one mentioned on #11066. This is the only case of this
combination in all of our regressions, which includes rerunning all
regressions through the `dump` tester (there are 9 cases where `let` and
variable shadowing are used in a benign way).

The user can always use `--fresh-binders` if they wish for cvc5 to be
SMT-LIB compliant, but this option is incompatible with proofs.

Fixes https://github.com/cvc5/cvc5-wishues/issues/152.