Fix strategy for datatypes solver involving acyclicity (#11094)
This fixes model soundness issues in the quantifier-free datatypes
solver. cvc5 could fail to recognize there is a conflict due to an
acyclicity inference. The issue occurs as early as cvc5 0.0.2.

The issue was caused by the strategy for the datatypes theory solver
using 2 separate loops for checking for cycles and instantiate (i.e.
inferring an equality based on a tester). The latter could introduce a
fact that should then lead to us checking need to check for cycles
again. However, we could terminate with "sat" if there was nothing else
to do. The solver has always had this structure, I suspect that a change
to the fact vs lemma policy made this design incorrect.

Adds a regression exhibiting the issue that is fixed now.