Change definition of quantifiers skolemize skolem to ensure closed index (#11071)
We require having indices be closed terms to ensure they can be
reparsed.

This reverts to a previous definition of `@quantifiers_skolemize` which
had this property.

When doing this change, a bug was discovered in the elimination of
witness terms. We require marking that the quantified formulas
corresponding to witness terms must preserve structure or else the
skolems are bogus. This PR adds infrastructure for doing this.

FYI @amaleewilson