Implement evaluation of semi-evaluated kinds in model (#11093)
This revises how APPLY_SELECTOR terms are evaluated in models.

Previously, APPLY_SELECTOR was marked "unevaluated" and hence a term `(s
x)` would only evaluate if `(s x)` could be rewritten or if `(s x)`
appeared syntactically in the model.

This updates `TheoryModel` so that APPLY_SELECTOR is marked
"semi-evaluated". It adds a further method to check whether the
evaluation of such a term is entailed by congruence, or else can be
assigned arbitrarily.

A similar change is made for `SEQ_NTH`, which had similar issues.

Fixes https://github.com/cvc5/cvc5/issues/10770.