Finalize renaming to ethos (#11109)
This updates the proof output of cvc5 to the new Eunoia syntax when
applicable.

It furthermore renames the signature files *.smt3 to *.eo and moves them
to the new appropriate directories.

The ethos version is update to one that generates a binary named ethos.

Further PRs will update the internal namings e.g. src/proofs/alf ->
src/proofs/eo, although this is lower priority.