Enable term-db-mode=relevant by default (#11112)
This option makes it so that only terms that appear in assertions are
considered e.g. for E-matching. This is in contrast to the default
behavior where additionally "preregistered" terms are considered.

This gives convincing results on SMT-LIB quantified logics:
+293-167 over 86550 benchmarks in logics with UF and quantifiers.
it is 9.7% faster on average on commonly solved instances 16820s vs
18467s.