Refine policy for lazy lambda lifting (#11140)
Certain lambdas may need to be lifted based on the symbols in their
body. In particular, a lambda like `(lambda ((x Int)) (+ 1 (f x))` may
need to be lifted since its body contains a free symbol `f`. In
particular equating `f` to this lambda leads to unsat, which we do via
lambda lifting.

This refines this policy so that the only lift such lambdas if they
become equal to other function symbols; otherwise their model value is
not relevant.

This also makes some minor fixes to the core model to ensure that lambda
values are fully reduced in models.