Update RARE reconstruction algorithm to allow free variables defined in conditions (#10914)
Allows RARE rules to contain free variables x not in the left hand side
of a rule, provided they occur defined in a condition e.g. an equality x
= t, where t contains only variables that are in the left hand side of
the rule.

This updates the README for RARE.

It also fixes an inconsistency between the evaluator and the rewriter
for `ints.log2` which was discovered while testing this change.