[alethe] Refactor translation of ASSUME/SCOPE (#11167)
With this change the Alethe post-processor uses `ProofRule::ASSUME`,
which would be helpful in a future effort to use the same printer as
CPC. This commit also refactors the translation of SCOPE.

Only a few commits are left to finish the migration of the fixes from
proof-new. This commit disables the few active regressions in main that
test the Alethe backend. They will be reactivated once the migration in
finished.