Fix issue with non-standard RE ranges in regular expression consume (#11174)
This bug accidentally reverses the contents of a regular expression
concatentation in certain range cases, which would lead to unsoundess if
we supported non-standard RE ranges.

This bug was introduced in the refactor here:
https://github.com/cvc5/cvc5/commit/20ffeae01dc565547015590e06efc4db0092734f,
which made the rewriter call the re consume utility in a new way is
sensitive to the order of the vector.

Fixes https://github.com/cvc5/cvc5/issues/11168.