Handle rare case of aliased purification variables in string infer proof cons (#11187)
We use a step that introduces purification skolems to ensure that
substitutions are applied properly to reason about string inferences.
However, this could lead to issues in very rare cases where the
purification skolem was already a part of the inference we are
processing.

This PR solves the issue by additionally purifying the purification
variables in the rare case this occurs. This occurs on 2 existing
regressions and in the regression added for the resolved issue.

Fixes https://github.com/cvc5/cvc5/issues/11164.