Add modes for assigning default values to functions in models (#11189)
Adds 3 modes for how models are constructed and implements them for the
first-order and higher-order cases. We maintain the default for first
order (the mode `first`).

Note this changes our default model construction for HO functions from
"first-enum" to "first" to be more consistent with what is done for
first-order.