[alethe] Enable Alethe tester in regressions (#11179)
The Alethe tester is amended to not fail when cvc5 does not support
producing an Alethe proof for a benchmark.

This commit also disables a number of regressions where it is not simple
for cvc5 to realize it cannot produce an Alethe proof. Also a few
scenarios where the Alethe translation has performance issues.

This commit can only go in after #11178 is merged.