bv-to-int: handle overflows directly (#11172)
Currently, the bit-vector overflow operators are eliminated by the BV
rewriter in all cases.

This PR conditions the elimination of most of them to the case where
int-blasting is disabled.
When int-blasting is enabled, most operators are not eliminated, but
instead are handled by the int-blaster.

The reason for this is that the translation to integers becomes much
more efficient on the operators themselves, compared to their rewritten
variants.
For example, if `xx` and `yy` are integers that correspond to `x` and
`y` (resp.), then `bvuaddo x y` simply translates to `x + y < 2^k`,
where `k` is the bit-width of `x` and `y`.

The only two exceptions are bvnego and bvsdivo, whose elimination in the
BV rewriter creates a simple equality.

---------

Co-authored-by: Yoni Zohar <yonizohar@Yonis-MacBook-Pro.local>
Co-authored-by: Mathias Preiner <mathias.preiner@gmail.com>