Make RARE rewrites static in CPC signature (#11193)
This makes it so that rewrite rules are defined statically in an *.eo
file instead of included on demand.

Adds a contrib script that auto-generates this file (along with what is
already necessary for updating the cvc5 API include).

In the future Ethos will make `declare-rule` unavailable in cpc proof
files.

NOTE: This may add some overhead since we have to parse the rewrites
*.eo file every time we run Ethos+CPC. We should run some performance
testing.