Fix bug in proof CNF stream (#11239)
This could cause us to mislabel an input clause as a theory lemma, due
to CNF stream being reentrant.

The fix follows a style used in the CNF stream already
https://github.com/cvc5/cvc5/blob/main/src/prop/cnf_stream.cpp#L210

This could potentially cause us to give unsound unsat cores when we are
proof producing, although we have never observed this.