Do not check for shadowed variables with --wf-checking (#11260)
Shadowed variables are now always allowed in our API.

This relaxes an API check for shadowed variables with `--wf-checking`
which is inconsistent with other behaviors of cvc5.

A similar relaxation was done to our assertions utility on
https://github.com/cvc5/cvc5/pull/10556. The same change should have
been applied for the code changed in this PR.

This also makes our implementation of `expr::hasFreeVar` robust to
variable shadowing.