Simplifications and improvements for the CPC printer (#11242)
This is towards a proof logging mode, which will print proofs of
preprocessing, theory lemmas and SAT refutations as they are learned.

This PR adds the necessary functionality to the CPC (ALF) printer for
this.

First, it makes a major simplification to the way quantified variables
are treated. We now always convert variables `x` of type `T` to
`(eo::var x T)` and then rely on the letification to handle them in a
uniform way. This means that we don't need any custom tracking of
variables in our print channel utility.

Second, we add support for a `printNext` function which prints the
"next" proof in a proof log. This requires keeping the let bindings as
data members of the printer. This also requires keeping a cache of proof
nodes we have already printed.

Third, we add support for different "scope" modes.