parser: Introduce parsing mode to configure level of strictness. (#11261)
This introduces option `--parsing-mode` to allow configuration of both
more strict but also more lenient parsing than default. Previous option
`--strict-parsing` is now configured to be an alias (via an options
handler) for `--parsing-mode=strict`.

Further, we now allow to parse symbols that start with `.` or `@`, which
we use for internally freshly introduced symbols, in lenient parsing
mode.