Downgrade the required use of fresh-binder to a warning (#11256)
This makes us introduce a fresh variable when fresh-binders is false in
the rare case where we have a let-term that induces a capturing
substitution. This happens for inputs like `(forall ((x Int)) (let ((y
x)) (forall ((x Int)) (P y))`.

This behavior was introduced in https://github.com/cvc5/cvc5/pull/11079,
and actually causes several errors to be caught and thrown in SMT-LIB,
which are now warnings after this PR.

As a consequence of this PR, when this warning is thrown:
- Proof reference checking will fail unless this behavior is modelled
upstream in the proof checkers,
- The parser is not deterministic since it resorts to constructing a
fresh variable for the inner quantified formula.

Note that the latter does not really impact applications e.g.
distributed solving which rely on sharing quantified lemmas since the
quantified formulas we share will have resolved issues with variable
shadowing.

---------

Co-authored-by: Aina Niemetz <aina.niemetz@gmail.com>