Remove special handling of proofs in PropPfManager (#11254)
We have decided that SAT solvers should be responsible for managing
their proofs internally.

Thus, there is no need for custom handling in the PropPfManager.

This PR deletes the custom code that was necessary for generating proofs
given the limitations of older versions of CaDiCaL.

After this PR, we should revisit proof support in the CaDiCaL solver
based on various prop proof modes (e.g. drat, lrat, sat-external-prove).