[proofs] [alethe] Improve error flagging to catch unsupported types (#11235)
This allows cvc5 to realize it is in an unsupported fragment of Alethe
also when the problem contains only variables of unsupported theories,
which requires us to look at their type. As a consequence, we can
re-enable a few regressions.

This fixes the nightlies, where a build with Cocoa (which is not tested
together with Alethe in CI) was generating proofs without FF operators
but with FF variables.