[proofs] [alethe] Fix translations of CONTRA rule
Rules translated via resolution of unprocessed premises need to be
post-processed to reason about whether the premises are singleton clauses or
not. This is marked by the use of the AletheRule::RESOLUTION_OR rule in the
Alethe post-processor. The translation of ProofRule::CONTRA was not using the
proper resolution, which could lead to invalid `resolution` steps in the final
Alethe proof. This commit fixes this.