Add rational parsing to RARE (#11282)
Also adds a several initial rewrites that are possible now, and some
basic support for transcendental elimination rules.

Note this PR also makes it so that `cvc5_proof_rule.h` is *not*
automatically clang-formatted, since it breaks the documentation in that
include file.

This PR also fixes our printer for RARE rules in Eunoia and removes
several rules that are subsumed by ARITH_POLY_NORM and had usages of `(-
1)` which were not being translated to Eunoia properly.

This furthermore fixes an inconsistency with `printArithLitToken` which
would print decimals for integral rationals whereas non-integral
rationals would be printed as rationals; now both are printed as
rationals.

This fills in 29 trusted proof steps in our regressions.