Minor change to TrustId::SUBS_EQ (#11289)
This makes a minor change to how trusted steps for substitution
equivalence (SUBS_EQ) are stored.

Instead of
```
t = s
------ trusted
x = r 
```
We store 
```
        -----------------trusted
t = s  (t=s) = (x=r)
-------------------------EQ_RESOLVE
 x=r
 ```

This makes a difference since in the latter case, the RARE reconstruction can be applied. In fact 15 of the 33 SUBS_EQ holes in our regressions are fillable by RARE after this change.