[proofs] [alethe] Print valid arguments in hole steps (#11287)
Before we were printing an undeclared identifier for an unknown rule or
trust id. This commit also updates the version of Carcara, which was not
parsing arguments of the `hole` rule before, which could lead to issues
if a shared term has a name introduced to it in the arguments of a
`hole` rule.