Make arith operator elimination proof producing (#11280)
This module eliminates extended arithmetic operators (e.g. division)
during preprocessing.

This fills 281 trusted theory preprocessing steps in our regressions.

The main changes are to make the core method of arith::OperatorElim
static and accessible via the proof checker and the solver
simultaneously. The proof rule ARITH_OP_ELIM_AXIOM (which previously
existed as a placeholder) is now implemented. This required several
simplifications to the code, and the deletion of an expert option.

A followup PR will formalize this rule in CPC+Eunoia.