Eliminate mixed arithmetic in proofs prior to running RARE reconstruction (#11308)
This makes it so `--proof-elim-subtypes` is run before running RARE
reconstruction and after macro elaboration.

This requires another visit pass of the proof to recollect trusted
steps.

This is clearly the better order of operations, since it allows us to
search for RARE rewrites for the santized version of rewrites instead of
trying to repair RARE rewrites for the unsanitized versions.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>