Update BV RARE rules to not use evaluatable terms in index positions (#10904)
Updates bv rules to not use symbolic terms for indices.

There were two rules `bv-zero-extend-ult-const-1` and
`bv-zero-extend-ult-const-2` where it was not clear how to do this
elimination since a symbolic index of an extract term was introduced in
a condition. Moreover, these rules are not covered by regressions.

I am deleting them for now until we are able to (1) rewrite them in a
way that avoids the issue (2) add regressions that cover them.