More changes for safe-options (#11290)
This is work towards having 100% proof coverage when safe-options is
enabled.

This PR makes the following changes:
(1) several options related to proofs are now common options, as they
should be considered part of our core infrastructure,
(2) a few components (sets+cardinality, relations and fmf-fun) are
demoted to expert,
(3) a few miscellaneous debugging options (stats and printing) are
changed to common. This is to support testing setups where these can be
used for testing cvc5 when safe-options is enabled.