Fixes and changes to BV overflow elimination rules (#11316)
Fixes `bv-uaddo-eliminate` which had an off by one error. Also refactors
`bv-ssubo-eliminate` and `bv-saddo-eliminate` to match the expected form
implemented by the rewriter.

Also makes a slight change to the preprocessing converter to RARE to
ensure ACI_NORM is only applied to STRING_CONCAT terms. Applying it to
AND/OR was making it difficult to apply `bv-ssubo-eliminate` and
`bv-saddo-eliminate` due to reordering AND/OR.