Add Eunoia definitions for missing quantifiers rules (#11305)
Note that `ProofRewriteRule::QUANT_MINISCOPE` does not currently work
for this Eunoia definition, since the internal implementation of this
rule includes some hidden alpha equivalence steps. A followup PR will
make the existing form of this a MACRO rule and then reconstruct the
proofs into one that replays in the Eunoia signature.

Disables cpc on a regression that involves an unhandled case of variable
shadowing.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>