Add missing `#include`s
Each of these is necessary for the header file to make sense in isolation.
Wherever this commit adds `#include $A` to `$B.h`, this means that in trnaslation units,
the includes happened to be ordered as
```
...
```

This change makes the translation units robust to reordering of the includes,
and permits the use of precompiled headers.

In some cases, the previous code made an attempt to forward-declare other classes,
but in fact the forward declaration was insufficient and only worked in the presence of a previously-`#included` full definition.

In more detail:

* Forward declaration insufficient:
  * `src/context/default_clean_up.h`: takes a `std::vector` by value
  * `src/proof/unsat_core.h`: holds `Node`s by value
  * `src/prop/registrar.h`: declares a function taking a `Node` by value
  * `src/theory/arith/linear/arithvar_node_map.h`: holds a `DenseMap` by value
  * `src/theory/bv/theory_bv_rewrite_rules_core.h`: stack-allocates `BitVector`
  * `src/theory/uf/equality_engine_types.h`: struct holds a `Node` field
* Missing forward-declarations:
  * `src/theory/arith/linear/constraint_forward.h`: needs `Rational`
  * `src/util/iand.h`: needs `ostream` overloads
  * `src/util/indexed_root_predicate.h`: needs `ostream` overloads
  * `src/util/safe_print.h`: needs `abort`

Signed-off-by: Eric Wieser <efw@google.com>