Fix getInterpolant when top-level substitutions are present (#11340)
This corrects a bug where our results for interpolation are wrong when
top-level substitutions are present.

In particular, since we apply top-level substitutions to the goal, these
equalities must be conjoined with the interpolant we find. This bug was
introduced here: https://github.com/cvc5/cvc5/pull/10252. However this
also uncovers a further restriction: the conjoined equalities must
satisfy the "shared variable" requirement of get-interpolant. This
further refines the interpolation solver by only taking substitutions
that satisfy this requirement, other top-level substitutions are
re-asserted.

Note that `--check-interpolants` also did not catch this issue since we
were checking that the interpolant was valid *after* applying top-level
substitutions. This PR also fixes this issue so that
`--check-interpolants` checks the validity of interpolants on the
original goal. A similar fix is applied for abduction (which however
notice does not have the same complications with top-level
substitutions).

Fixes https://github.com/cvc5/cvc5/issues/11332.