Make interface for ppStaticLearn able to track proofs (#11311)
Towards eliminating several major sources of holes in our regressions.

Note that staticly learned formulas are no longer conjoined to the
assertion where they are learned. Instead they are appended to the end.
This eases proof/unsat core tracking and is more in line with the other
preprocessing passes.

This should be performance tested.