Alethe: add name of trust rule as argument to alethe hole rules (#11398)
Currently, a trust rule is translated into an Alethe proof step that
looks like this:

`(step t95 (cl @p_34) :rule hole :args ("11" @p_34))`

i.e., the first argument is the trust rule id as a string but encoded as
an integer.

For debugging and evaluation it is helpful for me to get the name of a
trust rule instead, so for example:

`(step t95 (cl @p_34) :rule hole :args
("PREPROCESS_STATIC_LEARNING_LEMMA" @p_34))`

I only see advantages of this since the indices of the ids are cvc5
internal. If they are needed though maybe we could add both as an
argument?