Create self-contained JARs for cvc5 Java bindings (#11368)
This PR updates the cvc5 Java bindings to search for native libraries
within JARs in the classpath, in addition to the current search in
system directories and the paths specified in `java.library.path`. This
update enables the creation of self-contained JARs that include the cvc5
Java bindings along with the necessary native libraries. This is the
first step toward publishing self-contained JARs to Maven Central. For
now, the CI process creates a JAR for each platform and stores it under
the latest tag or the version being released.