Make QUANT_MERGE_PRENEX not eliminate duplicate variables, update QUANT_UNUSED_VARS (#11400)
Currently, the Eunoia definition of quant_merge_prenex is inaccurate
since it does not handle duplicate variables.

Introduces a new macro for the current behavior and reconstruction to
the version that does not eliminate duplicates, making the Eunoia
definition accurate.

This also updates the Eunoia definition of quant_unused_vars to handle
duplicates.

This should be merged before https://github.com/cvc5/cvc5/pull/11391, as
that PR had an alternative fix which will be subsumed by this approach.