Add generic macro for rewriting body of quantified formulas (#11391)
The quantifiers rewriter applies a set of rewrites recursively to bodies
of quantified formulas.

This introduces a macro for this method and fine grained proof
reconstruction for it. It introduces 2 new RARE rewrites for this
purpose. FYI @Lachnitt

It also makes a few minor fixes and refactorings in the relevant
classes.

Fills 27/91 remaining quantifier rewrite holes on regressions.

Also fixes a bug in the Eunoia signature for QUANT_MERGE_PRENEX, which
wasn't dropping duplicate variables.