Make several methods in NodeManager static (#11403)
This makes some of the methods in NodeManager that have explicit access
to a Node or TypeNode static.

This has two advantages:
(1) we don't need explict access to a `NodeManager*` where these methods
are called,
(2) for methods with a single argument, we guarantee the correct
NodeManager is used automatically, e.g. `NodeManager::mkBoundVar` always
uses the node manager associated with the given type.

This adds `TypeNode::getNodeManager` as a *protected* method, which can
be accessed by `NodeManager` already since it is a friend.

It cleans up some of the unecessary calls to `NodeManager::currentNM()`
at a few call sites.