Add proper interface for skolem information in Node (#11404)
Currently, to know information about what kind of skolem a Node is
requires manually looking up the appropriate `SkolemManager`.

This PR adds a proper interface to get skolem information from a `Node`.

This will be useful when eliminating calls to `NodeManager::currentNM`,
as this information no longer needs to be accessed manually.

This PR cleans up a few low-level instances of this.