Refactoring set defaults for safe options, full strict proofs (#11407)
This refactors setDefaults to disable nlCov if safe options is enabled,
and to enable "full strict" proofs.

To simplify the logic, nlCov and ufSymmetryBreaker are now *true* by
default and can be disabled. This allows us to have disabling these
options indicate that further changes should not reenable them.

It also ensures certain configurations of CEGQI are enabled with full
strict proofs.