Add ProofRule::ARITH_MULT_ABS_COMPARISON (#11297)
This adds proofs for a commonly used nonlinear lemma schema that
compares the absolute value of non-linear terms.

It does some minor refactoring of the nonlinear solver (e.g. to
eliminate use of mixed arithmetic) to setup these proofs and uses a new
rule `ProofRule::ARITH_MULT_ABS_COMPARISON`. As part of the
reconstruction for this rule, we require 3 new RARE rewrites. FYI
@Lachnitt

This appears in 69 lemmas in our regressions.

A followup PR will add the Eunoia definition for this rule.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>