Add macro and elaboration for quantifiers prenex (#11421)
This adds support for reconstruction of proofs of prenexing. This relies
on miniscoping (in reverse), thus no new proof rule is necessary.

The reconstruction does not always work, which may lead to further proof
holes, which will be addressed in future PRs.

A simple example is added which now has a complete proof.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>