Support empty set and sequence in RARE (#11348)
This PR will avoid having to introduce numerous ad-hoc rewrites for
sequences and sets.

Adds 3 new internal-only kinds for this purpose:
- Kind::TYPE_OF which represents the type of its argument
- Kind::SET_EMPTY_OF_TYPE, which is the empty set for the type
represented by its argument,
- Kind::SEQ_EMPTY_OF_TYPE, which is the empty sequence (or string) for
the type represented by its argument.

Each of these kinds are supported in the RARE parser. They are treated
as macros by RARE, that is, when RARE rules are instantiated, they are
immediately evaluated, so that e.g. `(@seq.empty_of_type (@type_of a)) =
(as set.empty (Set Int))` if `a` is a constant of type `(Set Int)`.

These operators are not exported to Eunoia. Instead, the list node
converter (which prints RARE rules) is updated to eliminate these
symbols during printing.

Note that we require passing types explicitly as arguments wherever
`Kind::TYPE_OF` is used. Furthermore we require changing the definition
of `seq.empty` to account for strings in RARE rules. This introduces an
internal `@seq.empty` symbol to denote empty sequence and changes
`seq.empty` to be a program to construct the empty string or sequence.
This means the RARE rules in Eunoia can use the `seq.empty` syntax in
the expected way.

This PR then further generalizes the remaining string rules to
sequences, simplifies RARE sets rules, and adds 2 missing rules from
sets.

---------

Co-authored-by: Lachnitt <lachnitt@stanford.edu>