Make DT_CONS_EQ macro, elaborate to simpler version (#11420)
This fixes proof failures in datatypes involving the rule DT_CONS_EQ.

This rule currently involves a more complex implementation than what is
documented and expected.

This fixes the documentation, makes it a macro, and introduces 2 new
rules for its elaboration. We then use the simpler version in the
datatypes infer proof constructor.

The proof elaboration for this macro is not yet complete, but this fixes
the open proof errors involving this rule, included is an example
regression.