Only try to use basic rewriter to automatically reconstruct trust steps (#11388)
This part of the code attempts to cast trust steps as rewrites, with the
hope of making DSL rewrite rule elaboration easier.

This PR changes the code so that we only try the basic rewriter, not the
extended one.

Since the extended rewriter does not have fine-grained proof support,
reconstructing trust steps in it does not provide much benefit.

Furthermore, it is less overhead to compute basic rewrites.