Make TrustId::THEORY_INFERENCE theory specific (#11425)
This simplifies statistics so that THEORY_INFERENCE is not further
subcategorized by a theory id. There is no need to do this since there
is no generic architecture at the moment that uses theory inference in a
theory-agnostic way.

Instead we introduce specific trust ids for each theory as needed.