Fix bug in sygus solver for trivially infeasible conjectures (#11481)
Currently, the sygus solver may return solutions for infeasible
conjectures in a corner case where the conjecture is trivially
infeasible.

This bug was introduced recently when we introduced a heuristic for
removing unconstrained functions-to-synthesize.

It also adds a few more safe guards to ensure similar corner cases do
not occur.

This adds a regression demonstrating the issue.