Reorganize final proof callback (#11424)
This reorganizes when the final proof callback is run. The motivation is
to more accurately record statistics for our proof holes and secondly
for performance.

In particular, previously it was being run whenever we connected a proof
to input assertions, e.g. for unsat cores, which was spurious.

Instead, we now only run the callback when `--check-proofs` is enabled.

This decouples the final proof callback from the proof post processor.
We now call `ProofManager::checkFinalProof` when applicable.

Also adds a statistic for which rules are handled in the *.eo signature.
Furthermore, it guards the computation of statistics when not using
`--stats-internal`.