Initialize the rewriter based on safe-options (#11300)
Makes it so that the floating points, bags, separation logic and finite
field theories do not provide a rewriter when they are disabled due to
`--safe-options`.

Makes it so that sets does not rewrite experimental arithmetic, set
cardinality or relation terms when their corresponding extensions are
disabled. Also makes arith operator elimination not reduce experimental
arithmetic terms when `arith-exp` is false.