Make arith static learning compatible with proofs and unsat cores (#11382)
This modifies the arith static learner to generate globally valid
assertions, thus allowing it to be used with proofs.

This removes a case where auto-configuration of options silently
disables a (non-expert) option when proofs are enabled.

It also sets up the infrastructure for providing proofs for this
utility, for now a TrustId.