Fix possible invalid model construction in datatypes (#11426)
This makes a fix to model construction in datatypes involving irrelevant
terms.

This error can only happen for `--theoryof-mode=term`, used in
quantified logics, e.g. when building candidate models.

This happens on one (unsolved) benchmark in SMT-LIB UFDT, a minimized
version of this benchmark is added as a regression (which we answer
unknown on).