Remove nested rewrite calls during preRewrite from arrays rewriter (#11402)
Invoking the rewriter as an oracle at preRewrite leads to several
unintuitive proof holes from THEORY_ARRAYS, where we are reasoning about
the equivalence of unrewritten ITE terms.

It also seems to be bad from a performance perspective, as it is making
a nested call to the rewriter (e.g. for checking equality of indices).