Make RE derive conflicts an expert option (#11504)
The inference `STRINGS_RE_DERIVE` uses legacy code for deriving
conflicts based on computing derivatives of regular expressions.

These kinds of conflicts are used in 2 regressions and never in SMT-LIB
unsat proofs, since there are other inferences that take priority.

This disables this inference by default and makes it an expert option to
enable it, for the sake of simplifying proofs.