Eliminate nested call to rewriter for beta reduction, generalize `ProofRule::BETA_REDUCE` (#11329)
This PR does two things:
(1) 

For handling beta reductions for lambdas that are not already rewritten,
previously we would rewrite `(APPLY_UF (lambda x P) t)` ---> `(APPLY_UF
r t)` where `r` was the result of a nested call to the rewriter for
`(lambda x P)`. This is required since the rewriter does not rewrite
inside of operators. However this made it impossible to reconstruct
proofs in general.

We now rewrite `(APPLY_UF (lambda x P) t)` ---> `(HO_APPLY (lambda x P)
t)` for such cases. This allows the lambda to be rewritten naturally and
this step is easy to justify.

The caveat now is that `HO_APPLY` may be introduced for non-HO logics.
We add a case to `ppRewrite` to convert any such fully applied
`HO_APPLY` term back to `APPLY_UF`. The guarding cases in the UF solver
are changed to accomodate this.

(2) 
Generalizes `ProofRule::BETA_REDUCE` to handle (single argument) beta
reduction for `HO_APPLY`.