Add 3 macro theory rewrites and a theory rewrite for sets (#11491)
The macro theory rewrites include a rule that is analogous to the rule
MACRO_ARRAYS_DISTINCT_ARRAYS rule. Recall this rule is non-trivial since
it is difficult to establish a type independent criteria for
"valueness".

This adds 2 additional rules for evaluating intersection and different,
and then a theory rewrite for "normalizing" union. The latter does not
require reasoning about the valueness of elements whereas the former 2
do, and thus will require further elaboration.