Add macro rule for capture avoidance in beta redexes and its elaboration (#11513)
This adds a macro rewrite to capture some of the rewrites in the UF
rewriter and adds proof elaboration for it, which is expected to be
accomplished by (possibly multiple steps of) alpha equivalence.

Note that it uses a more elegant solution than
https://github.com/cvc5/cvc5/pull/11329, which lifted APPLY_UF to
higher-order applications to ensure shadowing was eliminated. Now, the
rewriting the APPLY_UF is done "in-place", i.e. `(APPLY_UF L ts) --->
(APPLY_UF L' ts)`. As a result, we revert to a behavior where HO_APPLY
is never introduced in first order logics, making the change to TheoryUF
unecessary. It also means we don't rely on higher-order beta-reduction
in first-order logics. This reenables `alethe` on these regressions.

This rewrite is used to ensure that beta reduction does not introduce
capture avoiding substitutions. The UF rewriter is refactored to use the
reference implementation.

Adds a new utility method (getMatchConds) to node_algorith.h which will
be used for a few other pending tasks in proof elaboration.

Fills 1 proof hole in our regressions and 287 proof holes on SMT-LIB.